Tuning T0/T1 (ie +/sigma0 trans) vs rhoJa
By analyzing the prob of success of a given measurement of \theta -> \tilde{\theta}, we saw that it was useful to be able to modulate T0/T1 to maximize the prob. of success. This modulation can be done via the incident energy of the electron, or equivalently rhoJa ~ 1/E. We know that:
i) letting T0/T1 -> is trivial by making E sufficiently large
ii) in general T0/T1 increases with lower energy (larger rhoJa) as the electron spends more "time" with the molecule. It usually approaches an asymptote. For example in the model32 case with D = J12 = 0, the asymptotic value is of order 1. Modulating J does not seem to change this much!
iii) in general for D > 0 (which makes the transition to T0 energetically favorable), T0/T1 is actually linear in rhoJa. The closer D is to 0, the greater the slope (this surely has a limit though)(this is because we want the transition to be favorable, but a big energy seperation is bad ). This behavior allows us to modulate T0/T1 with complete freedom !

Behavior (iii) is obviously of great utility. This rules out the Mn dimer because it has D<0 However at the moment Jz is messing with this (has to do with Jx != Jz making the transition not energy favorable). With Jx and Jz turned off get linear behavior back